modifications
This commit removes the PG4 directory and all of its contents, including the style.css, index.html, and script.js files.

It also modifies the index.html and style.css files in the TG4 directory.

Specifically, it adds a new div element with the class "box2" to the index.html file, and it adds a number of CSS properties to the .box1 and .box2 classes in the style.css file.

These properties include:

* transition-timing-function: cubic-bezier(0.075, 0.82, 0.165, 1)
* transition-timing-function: ease-in-out
* transition-property: width
* transition-duration: 2500ms
* transition-delay: 300ms

Additionally, it adds the following CSS properties to the .box2 class:

* width: 150px
* height: 150px
* background-color: rgb(255, 0, 179)
* border: 1px solid rgb(81, 255, 0)
* animation: func 2500ms ease-in-out 0.5s infinite alternate

Finally, it adds the following keyframes to the style.css file:

```
@keyframes func{
    from{ transform: translate(0%);}
    25% {rotate: 360deg;}
    50% { transform: translate(100%);}
    75% { background-color: white;}
    100%{ transform: translate(200%);}
}